Cleanup qvg
