+++ flashdreams/examples/run_causal_wan21.py
@@ -66,12 +66,6 @@ from einops import rearrange
 
 from flashdreams.core.distributed import init as distributed_init
 from flashdreams.infra.decoder.base import StreamingVideoDecoder
-from flashdreams.infra.diffusion.noise import (
-    load_initial_noise_rollout,
-    select_initial_noise_chunk,
-    select_temporal_initial_noise_chunk,
-    stack_initial_noise_chunks,
-)
 from flashdreams.recipes.wan.config.causal_wan21 import (
     CAUSAL_WAN21_CONFIG_BUILDERS,
     DEFAULT_VIDEO_HEIGHT,
@@ -164,48 +158,6 @@ def parse_args() -> argparse.Namespace:
         default=None,
         help="Optional suffix for output mp4/json names.",
     )
-    parser.add_argument(
-        "--initial_noise_path",
-        type=Path,
-        default=None,
-        help=(
-            "Optional torch Tensor path for explicit initial noise. Accepts "
-            "[num_chunks, B, L, D] patchified noise or [B, total_T, C, H, W]."
-        ),
-    )
-    parser.add_argument(
-        "--save_initial_noise_path",
-        type=Path,
-        default=None,
-        help=(
-            "Optional path to save the exact initial noise used by this run "
-            "in [B, total_T, C, H, W] layout."
-        ),
-    )
-    parser.add_argument(
-        "--text_embeddings_path",
-        type=Path,
-        default=None,
-        help=(
-            "Optional torch Tensor or {'prompt_embeds': Tensor} path to use "
-            "instead of FlashDreams text encoding."
-        ),
-    )
-    parser.add_argument(
-        "--renoise_noise_folder",
-        type=Path,
-        default=None,
-        help=(
-            "Optional folder containing official scheduler re-noise tensors "
-            "named <idx>_<chunk>_<step>.pt."
-        ),
-    )
-    parser.add_argument(
-        "--save_latents_path",
-        type=Path,
-        default=None,
-        help="Optional path to save clean latent rollout in [B, T, C, H, W].",
-    )
     return parser.parse_args()
 
 
@@ -225,131 +177,6 @@ def _resolve_prompt(
             )
         return prompts[prompt_index]
     return prompt_or_txt_path
-
-
-def _wan_unpatchified_noise_shape(pipeline: WanInferencePipeline) -> tuple[int, ...]:
-    transformer = pipeline.diffusion_model.transformer
-    transformer_config = transformer.config
-    channels = transformer_config.network.in_dim
-    if getattr(transformer_config, "concat_image_mask_to_latent", False):
-        channels -= 4 + 16
-    height = getattr(transformer, "_output_height", None)
-    width = getattr(transformer, "_output_width", None)
-    if height is None:
-        height = DEFAULT_VIDEO_HEIGHT // WAN_VAE_SPATIAL_COMPRESSION
-    if width is None:
-        width = DEFAULT_VIDEO_WIDTH // WAN_VAE_SPATIAL_COMPRESSION
-    return (
-        *transformer_config.batch_shape,
-        transformer_config.len_t,
-        channels,
-        height,
-        width,
-    )
-
-
-def _prepare_initial_noise(
-    *,
-    pipeline: WanInferencePipeline,
-    device: torch.device,
-    rollout: torch.Tensor | None,
-    autoregressive_index: int,
-    should_draw: bool,
-) -> tuple[torch.Tensor | None, torch.Tensor | None]:
-    """Return generation noise plus unpatchified noise for saving."""
-    noise_in_unpatchified_shape = (
-        pipeline.diffusion_model.config._noise_in_unpatchified_shape
-    )
-    if rollout is not None:
-        latent_shape = pipeline.diffusion_model.latent_shape
-        if rollout.ndim == len(latent_shape) + 1 or (
-            rollout.ndim == len(latent_shape) and len(latent_shape) == 5
-        ):
-            initial_noise = select_initial_noise_chunk(
-                rollout,
-                autoregressive_index=autoregressive_index,
-                latent_shape=latent_shape,
-            )
-            initial_noise = initial_noise.to(
-                device=device,
-                dtype=pipeline.diffusion_model.dtype,
-            )
-            if len(latent_shape) == 5:
-                save_noise = initial_noise
-            else:
-                save_noise = pipeline.diffusion_model.transformer.unpatchify_and_maybe_gather_cp(
-                    initial_noise
-                )
-            if noise_in_unpatchified_shape:
-                initial_noise = save_noise
-            return (
-                initial_noise.to(device=device, dtype=pipeline.diffusion_model.dtype),
-                save_noise,
-            )
-
-        unpatchified = select_temporal_initial_noise_chunk(
-            rollout,
-            autoregressive_index=autoregressive_index,
-            chunk_shape=_wan_unpatchified_noise_shape(pipeline),
-        )
-        unpatchified = unpatchified.to(
-            device=device,
-            dtype=pipeline.diffusion_model.dtype,
-        )
-        if noise_in_unpatchified_shape:
-            return unpatchified, unpatchified
-        initial_noise = pipeline.diffusion_model.transformer.patchify_and_maybe_split_cp(
-            unpatchified
-        )
-        return initial_noise, unpatchified
-
-    if not should_draw:
-        return None, None
-
-    if noise_in_unpatchified_shape:
-        initial_noise = torch.randn(
-            _wan_unpatchified_noise_shape(pipeline),
-            device=device,
-            dtype=pipeline.diffusion_model.dtype,
-            generator=pipeline.diffusion_model.rng,
-        )
-        save_noise = initial_noise
-    else:
-        initial_noise = torch.randn(
-            pipeline.diffusion_model.latent_shape,
-            device=device,
-            dtype=pipeline.diffusion_model.dtype,
-            generator=pipeline.diffusion_model.rng,
-        )
-        save_noise = pipeline.diffusion_model.transformer.unpatchify_and_maybe_gather_cp(
-            initial_noise
-        )
-    return initial_noise, save_noise
-
-
-def _load_text_embeddings(path: Path | None) -> torch.Tensor | None:
-    if path is None:
-        return None
-    payload = torch.load(path, map_location="cpu")
-    if isinstance(payload, dict):
-        payload = payload["prompt_embeds"]
-    if not isinstance(payload, torch.Tensor):
-        raise TypeError(f"text embeddings must be a Tensor, got {type(payload)}")
-    return payload
-
-
-def _load_renoise_replay(
-    folder: Path | None,
-    prompt_index: int = 0,
-) -> list[torch.Tensor] | None:
-    if folder is None:
-        return None
-    paths = sorted(folder.glob(f"{prompt_index}_*.pt"))
-    if not paths:
-        raise FileNotFoundError(
-            f"No scheduler re-noise replay tensors found in {folder}"
-        )
-    return [torch.load(path, map_location="cpu") for path in paths]
 
 
 def main() -> None:
@@ -386,16 +213,8 @@ def main() -> None:
             args.window_size_t
         )
     pipeline = pipeline_config.setup().to(device=device)
-    pipeline.diffusion_model.transformer.set_scheduler_renoise_replay(
-        _load_renoise_replay(args.renoise_noise_folder)
-    )
 
     assert isinstance(pipeline, WanInferencePipeline)
-    initial_noise_rollout = (
-        load_initial_noise_rollout(args.initial_noise_path)
-        if args.initial_noise_path is not None
-        else None
-    )
 
     # Per-rollout latent (H, W). For T2V we pass these explicitly; for
     # I2V the pipeline derives them from the resized first-frame pixels.
@@ -424,7 +243,6 @@ def main() -> None:
         image=image,
         height=latent_h,
         width=latent_w,
-        text_embeddings=_load_text_embeddings(args.text_embeddings_path),
     )
 
     torch.cuda.synchronize()
@@ -434,32 +252,11 @@ def main() -> None:
     # ---------------------------------------------------------------- rollout
     chunks: list[torch.Tensor] = []
     stats_history: list[dict[str, float]] = []
-    initial_noise_chunks: list[torch.Tensor] = []
-    latent_chunks: list[torch.Tensor] = []
     for i in range(args.total_blocks):
         num_frames = pipeline.get_num_output_frames(i)
         print(f"autoregressive_index: {i}, num_frames: {num_frames}")
-        initial_noise, save_noise = _prepare_initial_noise(
-            pipeline=pipeline,
-            device=device,
-            rollout=initial_noise_rollout,
-            autoregressive_index=i,
-            should_draw=args.save_initial_noise_path is not None,
-        )
 
-        if save_noise is not None and args.save_initial_noise_path is not None:
-            initial_noise_chunks.append(save_noise.detach().cpu())
-
-        video_chunk = pipeline.generate(i, cache, initial_noise=initial_noise)
-        if args.save_latents_path is not None:
-            assert cache.final_state is not None
-            latent_chunks.append(
-                pipeline.diffusion_model.transformer.unpatchify_and_maybe_gather_cp(
-                    cache.final_state.clean_latent
-                )
-                .detach()
-                .cpu()
-            )
+        video_chunk = pipeline.generate(i, cache)
         stats = pipeline.finalize(i, cache)
         if stats is not None:
             stats_history.append({"autoregressive_index": i, **stats})
@@ -490,19 +287,6 @@ def main() -> None:
             with open(stats_path, "w") as f:
                 json.dump(stats_history, f, indent=2)
             print(f"saved per-AR-step stats to {stats_path}")
-
-        if args.save_initial_noise_path is not None:
-            args.save_initial_noise_path.parent.mkdir(parents=True, exist_ok=True)
-            torch.save(
-                stack_initial_noise_chunks(initial_noise_chunks),
-                args.save_initial_noise_path,
-            )
-            print(f"saved initial noise to {args.save_initial_noise_path}")
-
-        if args.save_latents_path is not None:
-            args.save_latents_path.parent.mkdir(parents=True, exist_ok=True)
-            torch.save(torch.cat(latent_chunks, dim=-4), args.save_latents_path)
-            print(f"saved clean latents to {args.save_latents_path}")
 
     # Drop captured CUDA graphs / private mempools BEFORE NCCL teardown so
     # they don't hold workspace buffers across the destroy. Otherwise the
