+++ flashdreams/flashdreams/infra/diffusion/model/base.py
@@ -64,8 +64,11 @@ class DiffusionModelConfig(InstantiateConfig["DiffusionModel"]):
     ``0`` skips ``add_noise``."""
 
     _noise_in_unpatchified_shape: bool = False
-    """Debug only: Create the noise in the unpatchified shape (to align with the open-source codebases).
-    Note this will hurt performance so only use it for debugging."""
+    """Draw scheduler noise in unpatchified ``[B,T,C,H,W]`` layout.
+
+    Self-Forcing/QVG open-source inference uses this layout before patchify,
+    so Wan configs enable it for reproduction parity.
+    """
 
 
 class DiffusionModel(nn.Module, Generic[TransformerCacheT]):
@@ -148,7 +151,6 @@ class DiffusionModel(nn.Module, Generic[TransformerCacheT]):
         autoregressive_index: int,
         cache: TransformerCacheT,
         input: Any = None,
-        initial_noise: Tensor | None = None,
     ) -> tuple[Tensor, "DiffusionModel.FinalState[TransformerCacheT]"]:
         """Run the denoising loop for one AR step.
 
@@ -158,10 +160,6 @@ class DiffusionModel(nn.Module, Generic[TransformerCacheT]):
             input: Optional per-AR-step encoder output. Patchified here and
                 forwarded to ``predict_flow`` / ``postprocess_clean_latent``,
                 then stashed on the returned ``FinalState`` for ``finalize``.
-            initial_noise: Optional caller-provided denoising start tensor.
-                Defaults to a Gaussian draw from this model's RNG. This is
-                used only by reproduction/alignment harnesses that must share
-                the exact same latent noise across codebases.
 
         Returns:
             ``(clean_latent, final_state)``. ``clean_latent`` is unpatchified;
@@ -171,46 +169,29 @@ class DiffusionModel(nn.Module, Generic[TransformerCacheT]):
             input = self.transformer.patchify_and_maybe_split_cp(input)
         cache.start(autoregressive_index)
 
-        if initial_noise is None:
-            if self.config._noise_in_unpatchified_shape:
-                # The `self.latent_shape` is patchified. To align with OSS
-                # Self-Forcing/QVG, draw noise in unpatchified [B,T,C,H,W]
-                # layout and let the scheduler operate in that layout.
-                dummy_latent = torch.empty(
-                    self.latent_shape, device=self.device, dtype=self.dtype
-                )
-                dummy_latent = self.transformer.unpatchify_and_maybe_gather_cp(
-                    dummy_latent
-                )
-                initial_noise = torch.randn(
-                    dummy_latent.shape,
-                    device=self.device,
-                    dtype=self.dtype,
-                    generator=self.rng,
-                )
-            else:
-                initial_noise = torch.randn(
-                    self.latent_shape,
-                    device=self.device,
-                    dtype=self.dtype,
-                    generator=self.rng,
-                )
+        if self.config._noise_in_unpatchified_shape:
+            # The `self.latent_shape` is patchified. To align with OSS
+            # Self-Forcing/QVG, draw noise in unpatchified [B,T,C,H,W]
+            # layout and let the scheduler operate in that layout.
+            dummy_latent = torch.empty(
+                self.latent_shape, device=self.device, dtype=self.dtype
+            )
+            dummy_latent = self.transformer.unpatchify_and_maybe_gather_cp(
+                dummy_latent
+            )
+            initial_noise = torch.randn(
+                dummy_latent.shape,
+                device=self.device,
+                dtype=self.dtype,
+                generator=self.rng,
+            )
         else:
-            expected_shape = self.latent_shape
-            if self.config._noise_in_unpatchified_shape:
-                dummy_latent = torch.empty(
-                    self.latent_shape, device=self.device, dtype=self.dtype
-                )
-                expected_shape = tuple(
-                    self.transformer.unpatchify_and_maybe_gather_cp(dummy_latent).shape
-                )
-            if tuple(initial_noise.shape) != tuple(expected_shape):
-                raise ValueError(
-                    "initial_noise shape must match latent_shape: "
-                    f"got {tuple(initial_noise.shape)}, "
-                    f"expected {tuple(expected_shape)}"
-                )
-            initial_noise = initial_noise.to(device=self.device, dtype=self.dtype)
+            initial_noise = torch.randn(
+                self.latent_shape,
+                device=self.device,
+                dtype=self.dtype,
+                generator=self.rng,
+            )
 
         def predict_flow(noisy_latent: Tensor, timestep: Tensor) -> Tensor:
             if self.config._noise_in_unpatchified_shape:
@@ -228,17 +209,15 @@ class DiffusionModel(nn.Module, Generic[TransformerCacheT]):
             )  # patchified output
 
             if self.config._noise_in_unpatchified_shape:
-                # The we need to unpatchify it again, to make sure scheduler operates
+                # Then unpatchify it again so the scheduler operates
                 # on the unpatchified latent.
                 output = self.transformer.unpatchify_and_maybe_gather_cp(output)
             return output
 
-        renoise_noise_fn = getattr(self.transformer, "scheduler_renoise_noise", None)
         clean_latent = self.scheduler.sample(
             initial_noise=initial_noise,
             predict_flow=predict_flow,
             rng=self.rng,
-            renoise_noise_fn=renoise_noise_fn,
         )
 
         if self.config._noise_in_unpatchified_shape:
