+++ flashdreams/flashdreams/infra/pipeline/base.py
@@ -194,7 +194,6 @@ class StreamInferencePipeline(
             StreamingEncoderCacheT, TransformerCacheT, StreamingDecoderCacheT
         ],
         input: Any = None,
-        initial_noise: Tensor | None = None,
     ) -> Tensor:
         """Generate one chunk for this AR step.
 
@@ -206,8 +205,6 @@ class StreamInferencePipeline(
                 is configured, must be ``None`` otherwise. Use
                 ``NullEncoderConfig`` to pass an already-encoded tensor
                 straight through.
-            initial_noise: Optional denoising start tensor for alignment
-                harnesses. Defaults to the diffusion model's Gaussian draw.
 
         Returns:
             Decoded tensor (e.g. RGB video) when a decoder is configured;
@@ -246,7 +243,6 @@ class StreamInferencePipeline(
             autoregressive_index=autoregressive_index,
             cache=cache.transformer_cache,
             input=input,
-            initial_noise=initial_noise,
         )
         cache.final_state = final_state
