+++ flashdreams/flashdreams/recipes/wan/config/causal_wan21.py
@@ -22,7 +22,6 @@ from typing import Any, TypedDict
 
 from torch import Tensor
 
-from flashdreams.core.attention.kv_compress import KVCompressionConfig
 from flashdreams.infra.config import InstantiateConfig
 from flashdreams.infra.diffusion.model import DiffusionModelConfig
 from flashdreams.infra.diffusion.scheduler.fm import FlowMatchSchedulerConfig
@@ -135,8 +134,7 @@ def _transformer_config(
     compile_network: bool,
     len_t_latent: int = _DEFAULT_LEN_T_LATENT,
     stamp_image_latent: bool = False,
-    kv_compression: KVCompressionConfig | None = None,
-    scheduler_renoise_unpatchified: bool = False,
+    kv_compression: Any | None = None,
 ) -> Wan21TransformerConfig:
     """Wan 1.3B transformer defaults for causal/streaming inference."""
     return Wan21TransformerConfig(
@@ -152,31 +150,7 @@ def _transformer_config(
         sink_size_t=0,
         stamp_image_latent=stamp_image_latent,
         kv_compression=kv_compression,
-        scheduler_renoise_unpatchified=scheduler_renoise_unpatchified,
         compile_network=compile_network,
-    )
-
-
-def _qvg_kv_compression_config(*, quant_type: str) -> KVCompressionConfig:
-    """QVG defaults matching the open-source Self-Forcing scripts."""
-    return KVCompressionConfig(
-        backend="qvg",
-        schedule={"compress_every_n_chunks": 8},
-        protected_recent_chunks=0,
-        protected_sink_tokens=0,
-        backend_config={
-            "quant_type": quant_type,
-            "cache_num_k_centroids": 256,
-            "cache_num_v_centroids": 256,
-            "kmeans_max_iters": 2,
-            "quant_block_size": 64,
-            "num_prq_stages": 1,
-            "scale_dtype": "float8_e4m3fn",
-            "kmeans_init": "random",
-            "store_prerope_keys": True,
-            "kernel_impl": "official_triton",
-            "preserve_rng": False,
-        },
     )
 
 
@@ -212,7 +186,6 @@ def build_self_forcing(
             transformer=_transformer_config(
                 checkpoint_path=AVAILABLE_CAUSAL_WAN21_CHECKPOINT_PATHS["self_forcing"],
                 compile_network=compile_network,
-                scheduler_renoise_unpatchified=True,
             ),
             scheduler=_scheduler_config(num_inference_steps=4, shift=8.0),
         ),
@@ -237,75 +210,8 @@ def build_self_forcing_lighttae(
             transformer=_transformer_config(
                 checkpoint_path=AVAILABLE_CAUSAL_WAN21_CHECKPOINT_PATHS["self_forcing"],
                 compile_network=compile_network,
-                scheduler_renoise_unpatchified=True,
             ),
             scheduler=_scheduler_config(num_inference_steps=4),
-        ),
-    )
-
-
-def build_self_forcing_qvg_int2(
-    *,
-    cp_size: int = 1,
-    compile_network: bool = True,
-    seed: int = 42,
-    i2v: bool = False,
-    enable_sync_and_profile: bool = False,
-) -> WanInferencePipelineConfig:
-    """Self-Forcing distilled checkpoint with QVG INT2 KV compression."""
-    assert not i2v, "QVG v1 config currently supports T2V only"
-    assert cp_size == 1, "QVG v1 config currently supports single-GPU only"
-    return WanInferencePipelineConfig(
-        enable_sync_and_profile=enable_sync_and_profile,
-        encoder=None,
-        decoder=_wan_vae_decoder_config(),
-        diffusion_model=DiffusionModelConfig(
-            seed=seed,
-            _noise_in_unpatchified_shape=True,
-            transformer=_transformer_config(
-                checkpoint_path=AVAILABLE_CAUSAL_WAN21_CHECKPOINT_PATHS[
-                    "self_forcing"
-                ],
-                compile_network=compile_network,
-                scheduler_renoise_unpatchified=True,
-                kv_compression=_qvg_kv_compression_config(
-                    quant_type="triton-nstages-kmeans-int2"
-                ),
-            ),
-            scheduler=_scheduler_config(num_inference_steps=4, shift=8.0),
-        ),
-    )
-
-
-def build_self_forcing_qvg_int4(
-    *,
-    cp_size: int = 1,
-    compile_network: bool = True,
-    seed: int = 42,
-    i2v: bool = False,
-    enable_sync_and_profile: bool = False,
-) -> WanInferencePipelineConfig:
-    """Self-Forcing distilled checkpoint with QVG INT4 KV compression."""
-    assert not i2v, "QVG v1 config currently supports T2V only"
-    assert cp_size == 1, "QVG v1 config currently supports single-GPU only"
-    return WanInferencePipelineConfig(
-        enable_sync_and_profile=enable_sync_and_profile,
-        encoder=None,
-        decoder=_wan_vae_decoder_config(),
-        diffusion_model=DiffusionModelConfig(
-            seed=seed,
-            _noise_in_unpatchified_shape=True,
-            transformer=_transformer_config(
-                checkpoint_path=AVAILABLE_CAUSAL_WAN21_CHECKPOINT_PATHS[
-                    "self_forcing"
-                ],
-                compile_network=compile_network,
-                scheduler_renoise_unpatchified=True,
-                kv_compression=_qvg_kv_compression_config(
-                    quant_type="triton-nstages-kmeans-int4"
-                ),
-            ),
-            scheduler=_scheduler_config(num_inference_steps=4, shift=8.0),
         ),
     )
 
@@ -367,8 +273,6 @@ def build_causal_forcing_framewise(
 CAUSAL_WAN21_CONFIG_BUILDERS: dict[str, Callable[..., WanInferencePipelineConfig]] = {
     "self_forcing": build_self_forcing,
     "self_forcing_lighttae": build_self_forcing_lighttae,
-    "self_forcing_qvg_int2": build_self_forcing_qvg_int2,
-    "self_forcing_qvg_int4": build_self_forcing_qvg_int4,
     "causal_forcing_chunkwise": build_causal_forcing_chunkwise,
     "causal_forcing_framewise": build_causal_forcing_framewise,
 }
